scripts/EnchantingMaterialTracker/requirements.txt
pyyaml
watchdog
jsonschema
pandas
reportlab
